Fix tellraw issue when console or datapacks use tellraw without slash
The Fabric Discord Link mod has the ability to enable/disable tellraw commands form propagating to the Discord; however, it doesn't currently account for the console or other sources such as datapacks calling tellraw without the preceeding slash. This commit simply adds an additional check to the command check to look for tellraw without the slash (in addition to the check with the slash). There is likely a more efficient way to write this code, but readability of the current solution is high.